+++ src/lib.ts
@@ -17,11 +17,6 @@ import initWasm, {
 import { arrayToHex, expect, headerToMap } from './utils';
 import type { ParsedTranscriptData, ProofData } from './types';
 
-import {
-  RemoteAttestation,
-  generateNonce,
-} from './remote-attestation/remote-attestation';
-
 let LOGGING_LEVEL: LoggingLevel = 'Info';
 
 function debug(...args: any[]) {
@@ -29,8 +24,38 @@ function debug(...args: any[]) {
     console.log('tlsn-js DEBUG', ...args);
   }
 }
-export type { RemoteAttestation };
-export { generateNonce };
+
+export interface RemoteAttestation {
+  protected: string;
+  payload: string;
+  signature: string;
+  certificate: string;
+  payload_object: Payload;
+}
+
+export interface Payload {
+  module_id: string;
+  timestamp: number;
+  digest: string;
+  pcrs: Map<number, string>;
+  certificate: Uint8Array;
+  cabundle: Uint8Array[];
+  public_key: Buffer;
+  user_data: Uint8Array | null;
+  nonce: string | null;
+}
+
+/**
+ * It generates a random nonce of length 40 using hexadecimal characters.
+ * This nonce is used to ensure the uniqueness of the attestation.
+ * @returns {string} The generated nonce.
+ */
+
+export function generateNonce() {
+  return Array.from({ length: 40 }, () =>
+    Math.floor(Math.random() * 16).toString(16),
+  ).join('');
+}
 
 export async function verify_attestation(
   remote_attestation_base64: string,
